Extends the lag tracking from
https://github.com/camunda/camunda/pull/58107 (which covered only
pending/in-flight snapshot replication) with tracking for
pending/in-flight log appends, which are now included in the total
replication lag measurement.

Log replication lag is first updated by each local append - then, for
each append request sent to a follower we track a
monotonically-increasing watermark (the cumulative number of log bytes
sent to the follower in the entire lifetime of RaftMemberContext). On
acknowledgement of an append we check the associated watermark against
the current value - this lets us account for any previously replicated
appends that weren't yet acknowledged as well as guarding against any
stale callbacks.

We do need to interrogate the log reader for the outstanding bytes in
some cases, but not on the hot path - only when installing a snapshot or
resetting state.

Closes #56755.

+++ zeebe/atomix/cluster/src/main/java/io/atomix/raft/metrics/LeaderAppenderMetrics.java
@@ -63,7 +63,7 @@ public class LeaderAppenderMetrics extends RaftMetrics implements CloseableSilen
   }
 
   public void observeAppend(
-      final String memberId, final int appendedEntries, final int appendedBytes) {
+      final String memberId, final int appendedEntries, final long appendedBytes) {
     getAppendRate(memberId).increment(appendedEntries);
     getAppendDataRate(memberId).increment(appendedBytes / 1024f);
   }
